refactor(backend): break down 1297-line main.rs into clean modules
BREAKING CHANGE: Major refactor for code quality

- Split monolithic main.rs into focused modules:
  * file_system.rs - File operations (read, write, delete, copy, etc.)
  * language_server.rs - LSP communication and process management
  * clipboard.rs - File clipboard operations (copy/cut/paste)
  * tree_sitter_parser.rs - Code parsing and symbol extraction
  * terminal.rs - Terminal session management (simplified)
  * search.rs - File search functionality (simplified)
  * utils.rs - Utility functions (app dirs, command execution)

- Reduced main.rs from 1297 lines to 66 lines
- Better separation of concerns and maintainability
- All functionality preserved with proper module exports
- Fixed compilation errors and borrow checker issues

Clean code matters! 🧹
