+++ src-tauri/src/terminal.rs
@@ -1,0 +1,46 @@
+use std::collections::HashMap;
+use std::sync::{Arc, Mutex};
+use std::io::Write;
+
+pub struct TerminalSession {
+    // Simplified for now
+    id: String,
+}
+
+pub type TerminalMap = Arc<Mutex<HashMap<String, TerminalSession>>>;
+
+#[tauri::command]
+pub async fn create_terminal_session(
+    session_id: String,
+    working_directory: Option<String>,
+) -> Result<String, String> {
+    // Simplified implementation
+    Ok(session_id)
+}
+
+#[tauri::command]
+pub async fn write_to_terminal(
+    session_id: String,
+    data: String,
+) -> Result<(), String> {
+    // Simplified implementation
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn close_terminal_session(
+    session_id: String,
+) -> Result<(), String> {
+    // Simplified implementation
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn resize_terminal(
+    session_id: String,
+    rows: u16,
+    cols: u16,
+) -> Result<(), String> {
+    // Simplified implementation
+    Ok(())
+}
